+++ NEWS.md
@@ -11,9 +11,11 @@
 * Added more informative error messages when constructing pedigree objects fails due to incorrect or non-standard sex coding. 
 * Expanded and refactored tests for various sex code and ID input scenarios, including comprehensive testing of input types, missing values, and configuration options. 
 * Add default point scaling as a function of pedigree size
+* Add non-default segment size scaling as a function of pedigree size
 * Set Sex key to be hidden by default in ggpedigree
 * Slight tweaks to GoT dataset demonstration to improve visualization
 * Remove reshape2 dependency by using base R functions
+
 
 # ggpedigree 1.0.0.1
 ## cran resubmission